[fix] move generator class to make folder
