+++ src/Commands/Make/ComponentViewMakeCommand.php
@@ -4,7 +4,6 @@ namespace Nwidart\Modules\Commands\Make;
 
 use Illuminate\Foundation\Inspiring;
 use Illuminate\Support\Str;
-use Nwidart\Modules\Commands\GeneratorCommand;
 use Nwidart\Modules\Support\Config\GenerateConfigReader;
 use Nwidart\Modules\Support\Stub;
 use Nwidart\Modules\Traits\ModuleCommandTrait;
