+++ src/Commands/Make/ModelMakeCommand.php
@@ -3,7 +3,6 @@
 namespace Nwidart\Modules\Commands\Make;
 
 use Illuminate\Support\Str;
-use Nwidart\Modules\Commands\GeneratorCommand;
 use Nwidart\Modules\Support\Config\GenerateConfigReader;
 use Nwidart\Modules\Support\Stub;
 use Nwidart\Modules\Traits\ModuleCommandTrait;
